+++ src/main/java/com/sashimi/course/presentation/api/response/PublicCourseResponse.java
@@ -7,7 +7,7 @@ import java.math.BigDecimal;
 public record PublicCourseResponse(
         String instructorName,
         String title,
-        BigDecimal price,
+        Long price,
         String thumbnail,
         int totalDuration,
         BigDecimal ratingAvg,
