+++ src/main/java/com/sashimi/credit/application/result/CreditBalanceResult.java
@@ -3,6 +3,6 @@ package com.sashimi.credit.application.result;
 import java.math.BigDecimal;
 
 public record CreditBalanceResult(
-        BigDecimal balance
+        Long balance
 ) {
 }
